feat(utils): enhance test automation with input validation and artifact verification (#5180)
## Problem
Test automation for utils/collect had several issues:
1. Tests failed due to incorrect execution order (pytest's alphabetical file collection ignored @pytest.mark.order across files)
2. No upfront input validation - playbook ran even with invalid/missing input files
3. Artifact verification checked old bundles instead of current test run outputs
4. Missing environment variable resolution for OMNIA_DATA_PATH and OMNIA_PROJECT_NAME

## Solution

### Test Execution Order (#496)
- Merged test_playbook.py and test_log_collector.py into single test_collect.py
- Ensures @pytest.mark.order() works reliably for all tests
- Tests now run in correct sequence: input validation → playbook execution → output verification

### Input Validation
- Added upfront validation for collect_pxe.yml before playbook execution
- Validates file existence, YAML structure, and functional group names
- Tests fail fast (0.88s) instead of running full playbook (47s) with invalid input
- Added test datasets for missing and invalid input scenarios

### Artifact Verification
- Added TEST_START_TIME to verify bundles are from current test run
- Prevents false positives from old bundles in output directory
- Validates bundle timestamps before content verification

### Environment Variable Resolution
- Updated run_validation.sh to export OMNIA_DATA_PATH and OMNIA_PROJECT_NAME
- Test functions now read these from target host's /etc/omnia/omnia.env
- Ensures consistency with deployed environment configuration

## Test Results
- Scenario 1 (Valid input): 16 passed, 0 failed ✅
- Scenario 2 (Missing input): 2 passed, 14 failed (expected) ✅
- Scenario 3 (Invalid input): 13 passed, 3 failed (expected) ✅
- Scenario 4 (Fresh output dir): 16 passed, 0 failed ✅

## Performance Improvement
- Invalid input detection: 47s → 0.88s (98% faster)

## Files Changed
- test/utils/fvt/collect/test_collect.py: Merged test file (pylint score: 9.13/10)
- test/utils/conftest.py: Added TEST_START_TIME fixture
- test/utils/library/functions/: Enhanced validation functions
- test/utils/datasets/test_data/: Added test datasets (no sensitive data)
- test/utils/README.md: Comprehensive documentation update
- Deleted: test_playbook.py, test_log_collector.py (merged into test_collect.py)

## Backward Compatibility
- No changes to playbook behavior or input file format
- Existing workflows continue to work unchanged
- Test automation is opt-in enhancement

Fixes #496, Issue #4849
