+++ test/utils/datasets/test_data/input/collect_pxe.yml
@@ -1,0 +1,34 @@
+# Copyright 2026 Dell Inc. or its subsidiaries. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+---
+# Kubernetes control plane nodes (x86_64)
+service_kube_control_plane_x86_64:
+
+# Kubernetes worker nodes (x86_64)
+service_kube_node_x86_64:
+
+# Slurm controller nodes (x86_64)
+slurm_control_node_x86_64:
+
+# Slurm compute nodes (x86_64)
+slurm_node_x86_64:
+
+# Slurm compute nodes (aarch64)
+slurm_node_aarch64:
+
+# Login nodes (x86_64)
+login_node_x86_64:
+
+# Login compiler nodes (aarch64)
+login_compiler_node_aarch64:
